# openapi.yaml
...
openapi: 3.0.3
info:
  title: 养老机构评估-复核-告知-费用 流程 API（虚构量表演示）
  version: 1.1.0
  description: 行政流程演示：量表评估 → 管理复核 → 家属告知（可靠回执与签收闭环）→ 费用生效。使用虚构量表 DEMO_ADL，不构成医疗诊断、护理分级依据或真实护理建议。家属告知由离线模拟的多渠道投递器异步回传受理/送达/失败/签收事件，不连接真实短信或外部服务。
servers:
  -
    url: /api
...
paths:
...
  /assessments/{caseId}/notification:
    get:
      tags:
        - notifications
      summary: 该案件全部告知记录（失败历史、未确认尝试均保留；状态含闭环投影）
      parameters:
        -
          name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 告知记录列表
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/NotificationRecord"
...
  /assessments/{caseId}/notification/timeline:
    get:
      tags:
        - deliveries
      summary: 时间线：告知创建 / 内容版本 / 投递 / 回执（含未生效留痕）/ 签收
      parameters:
        -
          name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 时间线
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Timeline"
  /assessments/{caseId}/notification/{notificationId}/refresh:
    post:
      tags:
        - deliveries
      summary: 内容刷新：按案件当前状态重新生成告知内容，内容变化时版本 +1（旧版本登记留痕）
      parameters:
        -
          name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        -
          name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RefreshContent"
      responses:
        201:
          description: 刷新结果（refreshed 表示是否产生新版本）
        404:
          description: 不存在
...
components:
  schemas:
...
    RefreshContent:
      type: object
      properties:
        reason:
          type: string
...
    NotificationRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - PENDING
            - DELIVERED
            - FAILED
            - SIGNED
        notifiableStatus:
          type: string
          enum:
            - CONFIRMED
            - UNCONFIRMED
        message:
          type: string
        contentVersion:
          type: integer
        failureReason:
          type: string
          nullable: true
        attempts:
          type: integer
        lastAttemptAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Timeline:
      type: object
      properties:
        caseId:
          type: string
          format: uuid
        items:
          type: array
          items:
            type: object
            properties:
              seq:
                type: integer
              at:
                type: string
                format: date-time
              type:
                type: string
                enum:
                  - NOTIFICATION_CREATED
                  - CONTENT_VERSION
                  - DELIVERY_CREATED
                  - RECEIPT_ACCEPTED
                  - RECEIPT_DELIVERED
                  - RECEIPT_FAILED
                  - RECEIPT_SIGNED
                  - SIGNATURE
              summary:
                type: string
              notificationId:
                type: string
              deliveryNo:
                type: string
              attemptNo:
                type: integer
              contentVersion:
                type: integer
              applied:
                type: boolean
              detail:
                type: object